+++ metacar_hybrid/sceneapi.py
@@ -83,15 +83,15 @@ class SceneAPI:
         return self._scene_static_data
 
     def main_loop(self):
-        """生成器，每次迭代返回 :class:`~metacar.models.SimCarMsg` 和图像帧，场景结束时退出。
+        """生成器，每次迭代返回 :class:`~metacar_hybrid.models.SimCarMsg` 和图像帧，场景结束时退出。
 
         此方法是一个生成器，每次迭代会返回当前的仿真车辆消息和摄像头图像帧。
         当场景结束或连接中断时，生成器会自动退出。
 
         :return: 元组 (sim_car_msg, frames)，其中:
 
-            - sim_car_msg: :class:`~metacar.models.SimCarMsg` 对象，包含车辆状态、传感器数据等信息
-            - frames: 当前相机视图的列表，每个元素为 :class:`~metacar.models.CameraFrame` 对象
+            - sim_car_msg: :class:`~metacar_hybrid.models.SimCarMsg` 对象，包含车辆状态、传感器数据等信息
+            - frames: 当前相机视图的列表，每个元素为 :class:`~metacar_hybrid.models.CameraFrame` 对象
         """
         # 先发送 code2，告知场景已经就绪
         self._model_socket.send(Code2(code=2), Code2)